a/tar: remove old tar-1.13 version for pkgtools
GNU tar 1.28 contains a command-line switch to keep symbolic links
to directories in place on extracting tar files.
The Slackware package systems relied on this behavior, so it carried
an old version 1.13 of tar along to solve this issue.
Now that upstream GNU tar supports this semantics again and pkgtools
have been adapted to make use of it, we can get rid of the old version
and its building instructions.