a/pkgtools: adapt installpkg to new tar
Version 1.27 of GNU tar comes with a command line option to support
the old behavior of not clobbering symlinks we relied on in the
Slackware pkgtools.
Adapt installpkg to make use of this new parameter if it is supported
by the installed tar and thus remove the need for the legacy 1.13
version we kept around for this purpose.
For compatibility we still support tar versions in between 1.14 and
1.27 (requiring an extra 1.13 binary).
tar -t was never affected by the undesired behavior, so there is no
need to patch its invocations.